refactor: preview 각 컴포넌트에 배치
