+++ app/src/main/java/nextstep/shoppingcart/ui/screen/component/BackAppBar.kt
@@ -5,6 +5,7 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -19,5 +20,18 @@ fun BackAppBar(
         },
         modifier = modifier,
         navigationIcon = navigationIcon
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun BackAppBarPreview() {
+    BackAppBar(
+        title = "상품 상세",
+        navigationIcon = {
+            BackIconButton(
+                onBackClick = { }
+            )
+        }
     )
 }
