+++ app/src/main/java/nextstep/shoppingcart/ui/screen/component/BackIconButton.kt
@@ -6,6 +6,7 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 
 @Composable
 fun BackIconButton(
@@ -18,4 +19,10 @@ fun BackIconButton(
     ) {
         Icon(Icons.Filled.ArrowBack, contentDescription = "Back Button")
     }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun BackIconButtonPreview() {
+    BackIconButton(onBackClick = { })
 }
