+++ app/src/main/java/nextstep/shoppingcart/ui/screen/component/CenterAppBar.kt
@@ -10,6 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.tooling.preview.Preview
 
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -34,5 +35,14 @@ fun CenterAppBar(
                 )
             }
         }
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun CenterAppBarPreview() {
+    CenterAppBar(
+        title = "상품 목록",
+        onClick = { }
     )
 }
