+++ app/src/main/java/nextstep/shoppingcart/ui/screen/component/ProductItem.kt
@@ -9,6 +9,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import coil3.compose.AsyncImage
 import nextstep.shoppingcart.ui.utils.formatter.DefaultMoneyFormatter
@@ -43,4 +44,14 @@ fun ProductItem(
             text = "${formatter.format(price)}원",
         )
     }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun ProductItemPreview() {
+    ProductItem(
+        imageUrl = "https://www.picsum.photos/200",
+        title = "상품 이름",
+        price = 10000,
+    )
 }
